requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
boto3==1.33.13
botocore==1.33.13
certifi==2023.11.17
charset-normalizer==2.1.1
click==8.1.7
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.105.0
fastdup==1.73
h11==0.14.0
httptools==0.6.1
idna==3.6
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
jmespath==1.0.1
jupyter_client==8.6.0
jupyter_core==5.5.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.2
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
pillow-heif==0.14.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.28.1
s3transfer==0.8.2
sentry-sdk==1.39.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
starlette==0.27.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.18
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.12
websockets==12.0
